+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.1.0
+
+* `Versions`
+  * `kotlin`: `1.4.21` -> `1.4.30`
+  * `serialization`: `1.0.1` -> `1.1.0-RC`
+  * `exposed`: `0.28.1` -> `0.29.1`
+  * `tgbotapi`: `0.30.10` -> `0.32.5`
+  * `microutils`: `0.4.11` -> `0.4.25`
+* `Plugin`
+  * New method `BehaviourContext#invoke`
+    * Old method `invoke` has been deprecated
+
 ## 0.0.5
 
 * `Versions`